Now Categories 'on hover' display to the Right, with more Category Home Updates
Now, Categories 'on hover' display to the Right, used to display to the left and got cut off.

Added all the subcategories for "Personal Care" from the old IBYNN to the nav bar, and also added it to the search page.

Also Added The Amazon, Target, Walmart logo to the main page, and changed the description